requirements.txt
numpy
pandas
pytest
requests
pyyaml
psutil
codecarbon
scipy